feat: allow custom shell via `vitest.shellType` option
