+++ src/constants.ts
@@ -6,7 +6,6 @@ export const minimumNodeVersion = '18.0.0'
 
 export const distDir = __dirname
 export const workerPath = resolve(__dirname, 'worker.js')
-export const debuggerPath = resolve(__dirname, 'debug.js')
 export const setupFilePath = resolve(__dirname, 'setupFile.mjs')
 
 export const configGlob = '**/*{vite,vitest}*.config*.{ts,js,mjs,cjs,cts,mts}'
